+++ packages/google-cloud-sql/test/gapic_sql_instances_service_v1.ts
@@ -2202,12 +2202,10 @@ describe('v1.SqlInstancesServiceClient', () => {
       const request = generateSampleMessage(
         new protos.google.cloud.sql.v1.SqlInstancesInsertRequest(),
       );
-      const defaultValue1 = getTypeDefaultValue(
-        '.google.cloud.sql.v1.SqlInstancesInsertRequest',
-        ['project'],
-      );
-      request.project = defaultValue1;
-      const expectedHeaderRequestParams = `project=${defaultValue1 ?? ''}`;
+      request.body = {};
+      // path template: {region=**}
+      request.body.region = 'value';
+      const expectedHeaderRequestParams = 'region=value';
       const expectedResponse = generateSampleMessage(
         new protos.google.cloud.sql.v1.Operation(),
       );
@@ -2235,12 +2233,10 @@ describe('v1.SqlInstancesServiceClient', () => {
       const request = generateSampleMessage(
         new protos.google.cloud.sql.v1.SqlInstancesInsertRequest(),
       );
-      const defaultValue1 = getTypeDefaultValue(
-        '.google.cloud.sql.v1.SqlInstancesInsertRequest',
-        ['project'],
-      );
-      request.project = defaultValue1;
-      const expectedHeaderRequestParams = `project=${defaultValue1 ?? ''}`;
+      request.body = {};
+      // path template: {region=**}
+      request.body.region = 'value';
+      const expectedHeaderRequestParams = 'region=value';
       const expectedResponse = generateSampleMessage(
         new protos.google.cloud.sql.v1.Operation(),
       );
@@ -2284,12 +2280,10 @@ describe('v1.SqlInstancesServiceClient', () => {
       const request = generateSampleMessage(
         new protos.google.cloud.sql.v1.SqlInstancesInsertRequest(),
       );
-      const defaultValue1 = getTypeDefaultValue(
-        '.google.cloud.sql.v1.SqlInstancesInsertRequest',
-        ['project'],
-      );
-      request.project = defaultValue1;
-      const expectedHeaderRequestParams = `project=${defaultValue1 ?? ''}`;
+      request.body = {};
+      // path template: {region=**}
+      request.body.region = 'value';
+      const expectedHeaderRequestParams = 'region=value';
       const expectedError = new Error('expected');
       client.innerApiCalls.insert = stubSimpleCall(undefined, expectedError);
       await assert.rejects(client.insert(request), expectedError);
@@ -2314,11 +2308,9 @@ describe('v1.SqlInstancesServiceClient', () => {
       const request = generateSampleMessage(
         new protos.google.cloud.sql.v1.SqlInstancesInsertRequest(),
       );
-      const defaultValue1 = getTypeDefaultValue(
-        '.google.cloud.sql.v1.SqlInstancesInsertRequest',
-        ['project'],
-      );
-      request.project = defaultValue1;
+      request.body = {};
+      // path template: {region=**}
+      request.body.region = 'value';
       const expectedError = new Error('The client has already been closed.');
       client.close().catch(err => {
         throw err;