+++ core/packages/gax/CHANGELOG.md
@@ -6,6 +6,24 @@
 [1]: https://www.npmjs.com/package/gax-nodejs?activeTab=versions
 
 
+## [6.5.0](https://github.com/googleapis/google-cloud-node/compare/google-gax-v6.4.0...google-gax-v6.5.0) (2026-09-22)
+
+
+### Features
+
+* Add o11y gRPC tracing ([#9267](https://github.com/googleapis/google-cloud-node/issues/9267)) ([ff39233](https://github.com/googleapis/google-cloud-node/commit/ff39233da87eca175d3104467fc756010e9af0a3))
+* **gax:** Report response status codes on traced calls ([#9344](https://github.com/googleapis/google-cloud-node/issues/9344)) ([800322d](https://github.com/googleapis/google-cloud-node/commit/800322d5d1d116b93c9a5365fc8dd591037d2836))
+* **o11y:** Let callers enable tracing with env var ([#9377](https://github.com/googleapis/google-cloud-node/issues/9377)) ([ed9d279](https://github.com/googleapis/google-cloud-node/commit/ed9d27902e3ec71f99aa86a22135f30555c72dcd))
+* Pass through fallback parameter for HTTP calls ([#9277](https://github.com/googleapis/google-cloud-node/issues/9277)) ([653de60](https://github.com/googleapis/google-cloud-node/commit/653de605e5954a0e378962e3f859f99f607dc6cd))
+
+
+### Bug Fixes
+
+* **gax:** Detect aborts and timeouts by error name ([6ce394b](https://github.com/googleapis/google-cloud-node/commit/6ce394b9c15301f12150af1123905f8918688c16))
+* **gax:** Enforce the call deadline on the REST transport ([#9334](https://github.com/googleapis/google-cloud-node/issues/9334)) ([ddafe50](https://github.com/googleapis/google-cloud-node/commit/ddafe50c2487a7527eb57f7194dc63028fa22975))
+* **gax:** Map REST fallback transport errors to gRPC status codes ([be36fd9](https://github.com/googleapis/google-cloud-node/commit/be36fd984f07656c83902e46869db605a3ec259a))
+* **gax:** Report transport failures as UNAVAILABLE, as gRPC does ([db76247](https://github.com/googleapis/google-cloud-node/commit/db76247673d12026348ab3ddd9a1e487a9e6468a))
+
 ## [6.4.0](https://github.com/googleapis/google-cloud-node/compare/google-gax-v6.3.0...google-gax-v6.4.0) (2026-09-15)